implement postgres+pgbouncer as a separate test
This is done solely so that CI builds don't take too long.
As a solution this is run as a separate datastore test
leveraging build tags

consistency tests are left out of scope for now
